outsourced nav bar completely and tried to make login as PopUp
Some bugs still remaining. For example when u try to login on the settings page, the slider will be shown over the pop Up. Log In on the main page you are still able to enter letters in the game. After sending from nothing happens. Check that!
Also what is with the Sign Up and forgot password page? Same logic as popUp oder going to a next page? Maybe the second idea is more secure. Chat tomorrow